requirements.txt
anyio==3.7.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.98.0
h11==0.14.0
idna==3.4
pydantic==1.10.9
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
azure-functions==1.15.0
